coins_core
Module enc

Contains simplified access to bech32 and base58check encoder/decoder for Bitcoin addresses. Also defines common encoder errors.

Enums

EncodingError

Errors that can be returned by the Bitcoin AddressEncoder.

Functions

decode_base58

Decodes base58check into a byte string. Returns a EncodingError::Bs58Decode if unsuccesful

decode_bech32

Decode a witness program from a bech32 string. Caller specifies an expected HRP. If a different HRP is found, returns WrongHrp.

encode_base58

Encodes a byte slice to base58check with the specified version byte.

encode_bech32

Encode a byte vector to bech32. This function expects v to be a witness program, and will return an UnknownScriptType if it does not meet the witness program format.

Type Aliases

EncodingResult

A simple result type alias